pyproject.toml
[tool.poetry]
name = "tinyshare"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false
authors = ["Felix-Ecole <felix.ecole@laposte.net>"]

[tool.poetry.dependencies]
python = "3.11.9"
sanic = "^23.12.1"
tortoise-orm = "^0.21.3"
dotmap = "^1.3.30"
wheezy-template = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"